Move HivePartialAggregationPushdown to com.facebook.presto.hive.rule package
